requirements.txt
numpy>=1.21.0
pillow>=8.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.6.0
